+++ src/app/components/auth/Auth.module.scss
@@ -13,17 +13,6 @@
   padding: 40px 13px;
 }
 
-.authForm {
-  display: flex;
-  flex-direction: column;
-  gap: 28px;
-  padding: 56px 0 32px;
-
-  button {
-    height: 48px;
-  }
-}
-
 .memberCheck {
   // 제공된 타이포그라피 스타일에 없음
   font-size: 16px;
